requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==25.1.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
discord.py==2.4.0
filelock==3.13.1
Flask==3.1.2
frozenlist==1.5.0
fsspec==2024.6.1
gunicorn==23.0.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
libretranslatepy==2.1.1
lxml==5.3.0
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
mysql-connector-python==9.4.0
networkx==3.3
nltk==3.9.1
numpy==2.2.2
packaging==25.0
pandas==2.3.2
pillow==11.0.0
propcache==0.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
setuptools==70.2.0
six==1.17.0
sympy==1.13.3
textblob==0.19.0
threadpoolctl==3.5.0
torch==2.8.0+cu126
torchvision==0.23.0+cu126
tqdm==4.67.1
translate==3.6.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.2
Werkzeug==3.1.3
yarl==1.18.3
